Create get_entity_ids.py
Adds a script to gather Synapse IDs for entities with a name that matches an input string. Matches are searched for within each entity contained in a list of Synapse IDs provided at runtime.

Design
- searches through a list of Synapse project IDs (though any Synapse ID could work) provided as '-l'
- identifies Synapse objects that match the input string provided as '-n'
-outputs a csv containing the input Synapse ID and the Synapse ID for the entity with a name that matches the input string

Primary use case is to recover Synapse IDs for all 'publications' folders contained in Synapse projects that were previously created for each grant supported by the MC2 Center, based on the Synapse IDs of the grant-based projects.