Create synapse_json_schema_bind.py
Add first version of script that binds a JSON schema to a Synapse entity. Note that schematic-derived JSON schemas are not inherently compatible with Synapse JSON services. Some elements of this script are hard coded for MC2-specific resources. This script does not add annotations to entities with bound schemas, which is being handled by a separate script in development