Create table_to_annotations.py
Adds a framework for accessing annotations in a Synapse table and applying them to a Synapse entity. This script assumes that the entity has previously had a schema bound to it.

Note: this script access annotations and attempts to apply them to the correct entity, but has not yet been used to apply annotations. There are some formatting issues in the JSON schemas generated by schematic that need to be accounted for and column names in Synapse tables must be adjusted to remove spaces, in order to serve as keys in the 'key:value' annotation pairs.