Update sync_datasets.py
Added logic to parse link and alias for Synapse-related patterns. Script will preferentially copy an existing Synapse link from the link column to the synapseLink column. If an appropriate link doesn't exist, the script will check if alias is a synId. If true, a new Synapse URL will be constructed and listed in synapseLink