Update sync_education.py
Added columns for CCKP table schema
Added code to remove unused columns
Added logic to parse link and alias for Synapse-related patterns. Script will preferentially check if alias is a synId. If true, a new Synapse URL will be constructed and listed in synapseLink. If the alias is not a synId, the script will check if the Resource Link is a Synapse URL and copy it to the synapseLink column, if true. Resource links will have the format [synId](URL) for alias match and [Link](URL) if a URL match