Capture publication DOI for datasets (#104)
* Add code to pull DOI from pubs

Add one of the associated publication DOIs to dataset entry if the dataset does not already have a DOI available.

* Extend version label

When issues arise with portal table syncs, the previous version name format was causing errors when the sync was re-run, since it was only dependent on the month, day, and year. This makes it more unique, to avoid this problem.