Database management updates (#109)
* Create merge_and_correct_manifests.py

Adds a script that automates transfer of corrected entries in "updated" manifest to "merged" manifest. Merges based on the primary key of the entry. The script will also populate empty cells with placeholder annotations for publications and datasets

* Update .gitignore

* Update build_datasets.py

Updated script that chunks files in datasets and creates new datasets as needed to collect all files in dataset entry. New datasets are added to an updated DSP.