Add primary and foreign key handling
In a schematic-based model, if 'primary_key' is listed in 'Properties', 'is_key' will be 'true'
If the primary key of another model is listed (form component_id, e.g., 'Study_id') in Properties, a triple will be created to document the linkage. This is used by graphing software to render diagrams.